Moved tax.h to pennies

+++ src/tax/tax.h
@@ -1,9 +1,10 @@
 #pragma once
 #include <string>
+#include <cstdint>
 
 class Tax{
     public:
         virtual ~Tax() = default;
-        virtual float calculateTax(float amount) const = 0;
+        virtual int64_t calculateTax(int64_t amount) const = 0;
         std::string name = "Tax";
 };
